Implemented getEnemy by id with corresponding unit test

+++ src/main/java/com/codecool/mhmm/stickman/DAO/EnemyDao.java
@@ -2,19 +2,18 @@ package com.codecool.mhmm.stickman.DAO;
 
 import com.codecool.mhmm.stickman.GameObjects.Characters.Enemy.Enemy;
 import com.codecool.mhmm.stickman.GameObjects.GameObjectType;
-import com.codecool.mhmm.stickman.Map.Level;
 
 public interface EnemyDao {
 
     /**
      * get every field of enemy
      */
-    Enemy getEnemy(int id);
+    Enemy getEnemy(long id);
 
     /**
      * update specific enemy with this id
      */
-    void updateEnemy(Enemy enemy, int id);
+    void updateEnemy(Enemy enemy, long id);
 
     /**
      * update every enemy in this type
